FIX userlogin, created Token struct

+++ store/admin.go
@@ -10,9 +10,9 @@ func (us *UserStore) ListAll(offset int64, limit int64) ([]*model.User, error) {
 		// users       []Users
 		id          string
 		username    string
-		accesslevel int64
+		accesslevel int
 		tier        int
-		balance     int
+		balance     int64
 	)
 	q := `
 	SELECT id, username, accesslevel, tier, balance
